Splitted up functions to separated files Removed localtunnel support Added custom webhook backend support
