+++ functions/signOut.js
@@ -1,0 +1,7 @@
+module.exports = function () {
+    const platform = this;
+
+    // Clears the session information
+    platform.token = null;
+    platform.locations = [];
+}
