+++ package.json
@@ -11,8 +11,9 @@
     "nello.io"
   ],
   "dependencies": {
+    "localtunnel": "^1.9.0",
     "request": "^2.83.0",
-    "localtunnel": "^1.9.0"
+    "socket.io-client": "^2.1.1"
   },
   "repository": {
     "type": "git",
